feat: renamed file for unit testing
basic tests are in there are designed correctly // replacement "output" is not yet in there
